arp and bitwise shufflers. Still some debug messages on

+++ sequencer_utils.h
@@ -15,7 +15,7 @@ void char_binary_version_of_int(int num, char bin_num[17]);
 // void wchar_binary_version_of_pattern(seq_pattern p, wchar_t bin_num[17]);
 int pattern_as_int_representation(seq_pattern p);
 void wchar_binary_version_of_int(int num, wchar_t bin_num[17]);
-unsigned int gimme_a_bitwise_int(int pattern, int t);
+unsigned short gimme_a_bitwise_int(int pattern, int t);
 void convert_bitshift_pattern_to_pattern(int bitpattern, int *pattern_array,
                                          int len_pattern_array,
                                          unsigned gridsize);
